pyproject.toml
[tool.poetry]
name = "temporal-awareness-mcp"
version = "0.1.0"
description = "MCP server providing temporal awareness and time calculation capabilities for AI agents."
authors = ["pmbstyle <slava@trofimov.ca>"]
license = "MIT"
readme = "README.md"
packages = [{include = "temporal_awareness_mcp", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
mcp = "^1.2.0"
pydantic = ">=2.8.2"
python-dateutil = ">=2.9.0"
tzdata = "*"
starlette = "^0.39.0"
uvicorn = "^0.32.0"

[tool.poetry.scripts]
temporal-mcp-stdio = "temporal_awareness_mcp.stdio_main:main"
temporal-mcp-http = "temporal_awareness_mcp.http_main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
ruff = "^0.5.0"
httpx = "^0.28.1"
pytest-asyncio = "^0.23.7"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"